+++ CHANGELOG.md
@@ -36,6 +36,10 @@ Ref: https://keepachangelog.com/en/1.0.0/
 
 ## Unreleased
 
+### Features
+
+- [721](https://github.com/persistenceOne/pstake-native/pull/721) Add Query host chain user unbondings.
+
 ### Bug Fixes
 
 - [720](https://github.com/persistenceOne/pstake-native/pull/720) Fix unbondings loop.